Refactored Emila's changes to integrate with new file format
