+++ apps/algorithm/search/src/routes.py
@@ -4,9 +4,11 @@ from typing import Dict, Any
 from elasticsearch import Elasticsearch
 from elasticsearch.exceptions import NotFoundError
 from fastapi import APIRouter, HTTPException, Header
+
 from .config import DEFAULT_INDEX, DISTANCE_TO_SEARCH_WITHIN, ES_ENDPOINT
 from .database import insert_user_search
 from .enums import Status, SearchType, Sort
+from .models import Listing
 from .validation import validate_search_params
 
 search_router = APIRouter()
@@ -135,18 +137,30 @@ async def search(
 
 
 @search_router.post("/api/search/reindex/listing-created")
-async def reindex_listing_created(listingId: str, authorization: str = Header(None)):
-    # actual logic will go here
+async def post_listing(listing: Listing, authorization: str = Header(None)):
+    INDEX = os.getenv("ES_INDEX", DEFAULT_INDEX)
+    if (listing.price < 0):
+        raise HTTPException(status_code=422, detail="price cannot be negative")
+
+    es.index(index=INDEX, id=listing.listingId, body=listing.dict())
+
     return {"message": "Listing added successfully."}
 
 
 @search_router.patch("/api/search/reindex/listing-edited")
-async def reindex_listing_edited(listingId: str, authorization: str = Header(None)):
-    # actual logic will go here
+async def patch_listing(listing: Listing, authorization: str = Header(None)):
+    INDEX = os.getenv("ES_INDEX", DEFAULT_INDEX)
+    if (listing.price < 0):
+        raise HTTPException(status_code=422, detail="price cannot be negative")
+
+    es.index(index=INDEX, id=listing.listingId, body=listing.dict())
+
     return {"message": "Listing edited successfully."}
 
 
 @search_router.delete("/api/search/reindex/listing-deleted")
-async def reindex_listing_deleted(listingId: str, authorization: str = Header(None)):
-    # actual logic will go here
+async def delete_listing(listingId: str, authorization: str = Header(None)):
+    INDEX = os.getenv("ES_INDEX", DEFAULT_INDEX)
+    es.delete(index=INDEX, id=listingId)
+
     return {"message": "Listing deleted successfully."}
